그룹 생성 컴포넌트 분리 및 이름변경
1. Screen1 -> MainScreen -> GroupScreen 으로 이름변경
2. 그룹에서 공통으로 사용하는 그룹생성 버튼을 컴포넌트로 분리
